requirements.txt
python-telegram-bot==21.5
pillow==9.5.0
numpy==1.21.0
tensorflow==2.13.0
opencv-python-headless==4.5.5.64
requests==2.26.0
h5py==3.8.0
gdown



